move struct DynamicOptions to package pkg/server/dynamic_options.go
