+++ pkg/apiserver/authentication/oauth/options.go
@@ -17,11 +17,11 @@ limitations under the License.
 package oauth
 
 import (
-	"encoding/json"
 	"errors"
 	"net/url"
-	"strings"
 	"time"
+
+	"kubesphere.io/kubesphere/pkg/server/options"
 
 	"kubesphere.io/kubesphere/pkg/utils/sliceutil"
 )
@@ -91,57 +91,6 @@ type Options struct {
 	AccessTokenInactivityTimeout time.Duration `json:"accessTokenInactivityTimeout" yaml:"accessTokenInactivityTimeout"`
 }
 
-// DynamicOptions accept dynamic configuration, the type of key MUST be string
-type DynamicOptions map[string]interface{}
-
-func (o DynamicOptions) MarshalJSON() ([]byte, error) {
-	data, err := json.Marshal(desensitize(o))
-	return data, err
-}
-
-var (
-	sensitiveKeys = [...]string{"password", "secret"}
-)
-
-// isSensitiveData returns whether the input string contains sensitive information
-func isSensitiveData(key string) bool {
-	for _, v := range sensitiveKeys {
-		if strings.Contains(strings.ToLower(key), v) {
-			return true
-		}
-	}
-	return false
-}
-
-// desensitize returns the desensitized data
-func desensitize(data map[string]interface{}) map[string]interface{} {
-	output := make(map[string]interface{})
-	for k, v := range data {
-		if isSensitiveData(k) {
-			continue
-		}
-		switch v := v.(type) {
-		case map[interface{}]interface{}:
-			output[k] = desensitize(convert(v))
-		default:
-			output[k] = v
-		}
-	}
-	return output
-}
-
-// convert returns formatted data
-func convert(m map[interface{}]interface{}) map[string]interface{} {
-	output := make(map[string]interface{})
-	for k, v := range m {
-		switch k := k.(type) {
-		case string:
-			output[k] = v
-		}
-	}
-	return output
-}
-
 type IdentityProviderOptions struct {
 	// The provider name.
 	Name string `json:"name" yaml:"name"`
@@ -164,7 +113,7 @@ type IdentityProviderOptions struct {
 	Type string `json:"type" yaml:"type"`
 
 	// The options of identify provider
-	Provider DynamicOptions `json:"provider" yaml:"provider"`
+	Provider options.DynamicOptions `json:"provider" yaml:"provider"`
 }
 
 type Token struct {
