+++ pkg/simple/client/cache/cache.go
@@ -49,10 +49,6 @@ type Interface interface {
 	Expire(key string, duration time.Duration) error
 }
 
-// DynamicOptions the options of the cache. For redis, options key can be  "host", "port", "db", "password".
-// For InMemoryCache, options key can be "cleanupperiod"
-type DynamicOptions map[string]interface{}
-
 func RegisterCacheFactory(factory CacheFactory) {
 	cacheFactories[factory.Type()] = factory
 }
