+++ pkg/simple/client/cache/factory.go
@@ -1,8 +1,10 @@
 package cache
 
+import "kubesphere.io/kubesphere/pkg/server/options"
+
 type CacheFactory interface {
 	// Type unique type of the cache
 	Type() string
 	// Create relevant caches by type
-	Create(options DynamicOptions, stopCh <-chan struct{}) (Interface, error)
+	Create(options options.DynamicOptions, stopCh <-chan struct{}) (Interface, error)
 }
